new functionality was added according to final task of sprint 6

+++ src/service/epicService/AbstractEpicService.java
@@ -1,17 +1,18 @@
 package service.epicService;
 
-import java.util.ArrayList;
-
 import interfaces.HistoryManager;
 import interfaces.repository.EpicRepository;
 import interfaces.repository.SubtaskRepository;
 import interfaces.service.EpicService;
 import model.Epic;
-import model.EpicUpdationData;
 import model.Subtask;
 import model.TaskCreationData;
 import util.IdGenerator;
 import util.TaskManagerConfig;
+import util.TaskStatus;
+
+import java.util.ArrayList;
+import java.util.NoSuchElementException;
 
 public abstract class AbstractEpicService implements EpicService {
     private final EpicRepository epicRepo;
@@ -28,7 +29,13 @@ public abstract class AbstractEpicService implements EpicService {
 
     @Override
     public Epic create(TaskCreationData data) {
-        Epic epic = new Epic(idGenerator.generateNewId(), data);
+        if (data == null) {
+            data = new TaskCreationData(null, null);
+        }
+
+        Integer id = idGenerator.generateNewId();
+        Epic epic = new Epic(id, data);
+        epic.setStatus(TaskStatus.NEW);
 
         return epicRepo.create(epic);
     }
@@ -36,6 +43,11 @@ public abstract class AbstractEpicService implements EpicService {
     @Override
     public Epic get(Integer id) {
         Epic epic = epicRepo.get(id);
+
+        if (epic == null) {
+            throw new NoSuchElementException("epic not found");
+        }
+
         historyManager.add(epic);
         return epic;
     }
@@ -46,8 +58,16 @@ public abstract class AbstractEpicService implements EpicService {
     }
 
     @Override
-    public Epic update(EpicUpdationData data) {
-        Epic savedEpic = epicRepo.get(data.getId());
+    public Epic update(Integer id, TaskCreationData data) {
+        if (data == null) {
+            throw new IllegalArgumentException();
+        }
+
+        Epic savedEpic = epicRepo.get(id);
+
+        if (savedEpic == null) {
+            throw new NoSuchElementException("epic not found");
+        }
 
         savedEpic.setName(data.getName());
         savedEpic.setDescription(data.getDescription());
@@ -58,6 +78,10 @@ public abstract class AbstractEpicService implements EpicService {
     @Override
     public void remove(Integer id) {
         Epic epic = epicRepo.get(id);
+
+        if (epic == null) {
+            return;
+        }
 
         for (Subtask subtask : epic.getSubtasks()) {
             subtaskRepo.remove(subtask.getId());
@@ -75,6 +99,11 @@ public abstract class AbstractEpicService implements EpicService {
     @Override
     public ArrayList<Subtask> getSubtasks(Integer epicId) {
         Epic epic = epicRepo.get(epicId);
+
+        if (epic == null) {
+            throw new NoSuchElementException("epic not found");
+        }
+
         return epic.getSubtasks();
     }
 }
